feat: add slice in different file
