+++ src/redux/slices/index.ts
@@ -1,9 +1,7 @@
-import { ArticleDetailSlice } from "./articleDetails";
-import { articlesSlice } from "./articles";
+import ArticleDetailSlice from "./articleDetails";
+import articlesSlice from "./articles";
 
-// export const { reducer: articleReducer } = articlesSlice;
-// export const { reducer: articleDetailReducer } = ArticleDetailSlice;
 export default {
-  articles: articlesSlice.reducer,
-  articleDetail: ArticleDetailSlice.reducer,
+  articles: articlesSlice,
+  articleDetail: ArticleDetailSlice,
 };
